+++ src/main/java/com/sdl/selenium/utils/browsers/ChromeConfigReader.java
@@ -22,7 +22,7 @@ public class ChromeConfigReader extends AbstractBrowserConfigReader {
             "\n browser=chrome" +
             "\n browser.driver.path=src\\\\test\\\\resources\\\\drivers\\\\chromedriver.exe" +
             "\n browser.download.dir=src\\\\test\\\\resources\\\\download\\\\" +
-            "\n options.arguments=--lang=en --allow-running-insecure-content --enable-logging --v=1 --test-type --disable-extensions" +
+            "\n options.arguments=--lang=en --allow-running-insecure-content --enable-logging --v=1 --test-type" +
             "\n options.experimental.profile.default_content_setting_values.automatic_downloads=1" +
             "\n options.experimental.profile.content_settings.pattern_pairs.*.multiple-automatic-downloads=1" +
             "\n options.experimental.profile.default_content_settings.popups=0" +